It was only a test...this was just a test.
This was only a test.  This was just a test.  If this were more than just a test, then it would be MORE than just a test.  Do not panic.  Do not fuss.  Do not create a scene.  This is only  a test.


This was a test.  Go about your business.